backendgi/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
openpyxl==3.1.5
